+++ Options/SettingDefinitions/Pets.lua
@@ -152,8 +152,32 @@ class "SettingDefinitions.Pets" (function(_ENV)
     end
   end
   -----------------------------------------------------------------------------
-  --                        [Activity] Tab Builder                             --
+  --                        [Pet] Tab Builder                                --
   -----------------------------------------------------------------------------
+  function AcquireFixedColorSubOptions(self, parent, baseOrder, prefix, info)
+    local fixedColorPickerKey = prefix .. "FixedColorPicker"
+
+    if not self.PetTabControls[fixedColorPickerKey] then
+      local fixedColorInfo = info.fixedColor 
+
+      if fixedColorInfo then 
+        local fixedColorPicker = Widgets.SettingsColorPicker.Acquire(false, parent)
+        fixedColorPicker:SetID(baseOrder + 1)
+        fixedColorPicker:SetLabel(fixedColorInfo.label)
+        fixedColorPicker:BindUISetting(fixedColorInfo.setting)
+        self.PetTabControls[fixedColorPickerKey] = fixedColorPicker
+      end
+    end
+  end
+
+  function ReleaseNameFixedColorSubOptions(self, prefix)
+    local fixedColorPicker = self.PetTabControls[prefix.."FixedColorPicker"]
+    if fixedColorPicker then 
+      fixedColorPicker:Release()
+      self.PetTabControls[prefix.."FixedColorPicker"] = nil 
+    end
+  end
+
   function BuildPetTab(self)
     ---------------------------------------------------------------------------
     --- Background Section
@@ -168,14 +192,33 @@ class "SettingDefinitions.Pets" (function(_ENV)
     local showBackgroundCheckBox = Widgets.SettingsCheckBox.Acquire(false, backgroundSection)
     showBackgroundCheckBox:SetID(10)
     showBackgroundCheckBox:SetLabel(L.SHOW)
-    showBackgroundCheckBox:BindUISetting("activity.showBackground")
+    showBackgroundCheckBox:BindUISetting("pet.showBackground")
     self.PetTabControls.showBackgroundCheckBox = showBackgroundCheckBox
 
-    local backgroundColorPicker = Widgets.SettingsColorPicker.Acquire(false, backgroundSection)
-    backgroundColorPicker:SetID(20)
-    backgroundColorPicker:SetLabel(L.COLOR)
-    backgroundColorPicker:BindUISetting("activity.backgroundColor")
-    self.PetTabControls.backgroundColorPicker = backgroundColorPicker
+    local useFixedColorBackgroundCheckBox = Widgets.SettingsCheckBox.Acquire(false, backgroundSection)
+    useFixedColorBackgroundCheckBox:SetID(20)
+    useFixedColorBackgroundCheckBox:SetLabel(L.USE_FIXED_COLOR)
+    useFixedColorBackgroundCheckBox:BindUISetting("pet.useFixedColorForBackground")
+    self.PetTabControls.useFixedColorBackgroundCheckBox = useFixedColorBackgroundCheckBox
+
+    useFixedColorBackgroundCheckBox:SetUserHandler("OnCheckBoxClick", function(checkbox)
+      if checkbox:IsChecked() then 
+        self:AcquireFixedColorSubOptions(backgroundSection, 20, "background", {
+          fixedColor = { label = L.COLOR, setting = "pet.backgroundColor" }
+        })
+      else 
+        self:ReleaseNameFixedColorSubOptions("background") 
+      end 
+    
+    end)
+
+    if useFixedColorBackgroundCheckBox:IsChecked() then 
+      self:AcquireFixedColorSubOptions(backgroundSection, 30, "background", {
+        fixedColor = { label = L.COLOR, setting = "pet.backgroundColor" }
+      })
+    else  
+      self:ReleaseNameFixedColorSubOptions("background")
+    end
     ---------------------------------------------------------------------------
     --- Border Section
     ---------------------------------------------------------------------------
@@ -188,24 +231,88 @@ class "SettingDefinitions.Pets" (function(_ENV)
     local showBorderCheckBox = Widgets.SettingsCheckBox.Acquire(false, borderSection)
     showBorderCheckBox:SetID(10)
     showBorderCheckBox:SetLabel(L.SHOW)
-    showBorderCheckBox:BindUISetting("activity.showBorder")
+    showBorderCheckBox:BindUISetting("pet.showBorder")
     self.PetTabControls.showBorderCheckBox = showBorderCheckBox
 
-    local borderColorPicker = Widgets.SettingsColorPicker.Acquire(false, borderSection)
-    borderColorPicker:SetID(20)
-    borderColorPicker:SetLabel(L.COLOR)
-    borderColorPicker:BindUISetting("activity.borderColor")
-    self.PetTabControls.borderColorPicker = borderColorPicker
+
+    local useFixedColorBoderCheckBox = Widgets.SettingsCheckBox.Acquire(false, borderSection)
+    useFixedColorBoderCheckBox:SetID(20)
+    useFixedColorBoderCheckBox:SetLabel(L.USE_FIXED_COLOR)
+    useFixedColorBoderCheckBox:BindUISetting("pet.useFixedColorForBorder")
+    self.PetTabControls.useFixedColorBoderCheckBox = useFixedColorBoderCheckBox
+
+    useFixedColorBoderCheckBox:SetUserHandler("OnCheckBoxClick", function(checkbox)
+      if checkbox:IsChecked() then 
+        self:AcquireFixedColorSubOptions(borderSection, 20, "border", {
+          fixedColor = { label = L.COLOR, setting = "pet.borderColor" }
+        })
+      else 
+        self:ReleaseNameFixedColorSubOptions("border")
+      end
+    end)
+
+    if useFixedColorBoderCheckBox:IsChecked() then 
+      self:AcquireFixedColorSubOptions(borderSection, 30, "border", {
+        fixedColor = { label = L.COLOR, setting = "pet.borderColor" }
+      })
+    else  
+      self:ReleaseNameFixedColorSubOptions("border")
+    end
 
     local borderSizeSlider = Widgets.SettingsSlider.Acquire(false, borderSection)
     borderSizeSlider:SetID(30)
     borderSizeSlider:SetLabel(L.SIZE)
     borderSizeSlider:SetMinMaxValues(1, 10)
-    borderSizeSlider:BindUISetting("activity.borderSize")
+    borderSizeSlider:BindUISetting("pet.borderSize")
     self.PetTabControls.borderSizeSlider = borderSizeSlider
+    ---------------------------------------------------------------------------
+    --- Name Section
+    ---------------------------------------------------------------------------
+    local nameSection = Widgets.ExpandableSection.Acquire(false, self)
+    nameSection:SetExpanded(false)
+    nameSection:SetID(50)
+    nameSection:SetTitle(L.NAME)
+    self.PetTabControls.nameSection = nameSection
+
+    local namefont = Widgets.SettingsMediaFont.Acquire(false, nameSection)
+    namefont:SetID(10)
+    namefont:BindUISetting("pet.name.mediaFont")
+    self.PetTabControls.namefont = namefont
+
+    local nameTextTransform = Widgets.SettingsDropDown.Acquire(false, nameSection)
+    nameTextTransform:SetID(20)
+    nameTextTransform:SetLabel(L.TEXT_TRANSFORM)
+    nameTextTransform:SetEntries(TEXT_TRANSFORM_ENTRIES)
+    nameTextTransform:BindUISetting("pet.name.textTransform")
+    self.PetTabControls.textTransform = nameTextTransform
+
+    local nameUseFixedColorCheckBox = Widgets.SettingsCheckBox.Acquire(false, nameSection)
+    nameUseFixedColorCheckBox:SetID(30)
+    nameUseFixedColorCheckBox:SetLabel(L.USE_FIXED_COLOR)
+    nameUseFixedColorCheckBox:BindUISetting("pet.name.useFixedColor")
+    self.PetTabControls.nameUseFixedColorCheckBox = nameUseFixedColorCheckBox
+
+    nameUseFixedColorCheckBox:SetUserHandler("OnCheckBoxClick", function(checkbox)
+      if checkbox:IsChecked() then 
+        self:AcquireFixedColorSubOptions(nameSection, 30, "name", {
+          fixedColor = { label = L.TEXT_COLOR, setting = "pet.name.fixedColor" }
+        })
+      else 
+        self:ReleaseNameFixedColorSubOptions("name") 
+      end 
+    
+    end)
+
+    if nameUseFixedColorCheckBox:IsChecked() then 
+      self:AcquireFixedColorSubOptions(nameSection, 30, "name", {
+        fixedColor = { label = L.TEXT_COLOR, setting = "pet.name.fixedColor" }
+      })
+    else  
+      self:ReleaseNameFixedColorSubOptions("name")
+    end
   end
   -----------------------------------------------------------------------------
-  --                     [Activity] Release Builder                            --
+  --                        [Pet] Release Builder                            --
   -----------------------------------------------------------------------------
   function ReleasePetTab(self)
     for index, control in pairs(self.PetTabControls) do 
@@ -220,7 +327,7 @@ class "SettingDefinitions.Pets" (function(_ENV)
     local tabControl = Widgets.TabControl.Acquire(false, self)
     tabControl:SetID(1)
     tabControl:AddTabPage({
-      name = "General",
+      name = L.GENERAL,
       onAcquire = function() self:BuildGeneralTab() end,
       onRelease = function() self:ReleaseGeneralTab() end,
     })
@@ -232,7 +339,7 @@ class "SettingDefinitions.Pets" (function(_ENV)
     })
 
     tabControl:AddTabPage({
-      name = "Pets",
+      name = L.PET,
       onAcquire = function() self:BuildPetTab() end,
       onRelease = function() self:ReleasePetTab() end,
     })