+++ Retail/Contents/Pets/PetView.lua
@@ -15,6 +15,7 @@ export {
   FromBackdrop                        = Frame.FromBackdrop,
   RegisterUISetting                   = API.RegisterUISetting,
   FromUISetting                       = API.FromUISetting,
+  FromUISettings                      = API.FromUISettings
 }
 
 __UIElement__()
@@ -53,7 +54,6 @@ class "PetView" (function(_ENV)
   property "PetQuality" {
     type = Number
   }
-
 
   property "SpecieID" {
     type = Number
@@ -129,83 +129,109 @@ RegisterUISetting("pets.columns", 2)
 
 RegisterUISetting("pet.showBackground", true)
 RegisterUISetting("pet.showBorder", true)
+RegisterUISetting("pet.useFixedColorForBackground", false)
 RegisterUISetting("pet.backgroundColor", Color(35/255, 40/255, 46/255, 0.73))
+RegisterUISetting("pet.useFixedColorForBorder", true)
 RegisterUISetting("pet.borderColor", Color(0, 0, 0, 0.4))
 RegisterUISetting("pet.borderSize", 1)
 RegisterUISetting("pet.name.mediaFont", FontType("DejaVuSansCondensed Bold", 10))
 RegisterUISetting("pet.name.textTransform", "NONE")
+RegisterUISetting("pet.name.useFixedColor", false)
+RegisterUISetting("pet.name.fixedColor", nil)
+RegisterUISetting("pet.name.textColor", nil)
 -------------------------------------------------------------------------------
 --                              Observables                                  --
 -------------------------------------------------------------------------------
 function FromBackdropColor()
   local darkenFactor = 0.6
 
-  return FromUIProperty("PetQuality"):Map(function(quality)
-    local color = Color(ITEM_QUALITY_COLORS[quality - 1] or RED_FONT_COLOR)
-    return Color(color.r * darkenFactor, color.g * darkenFactor, color.b * darkenFactor, 0.73)
-  end)
+  return FromUIProperty("PetQuality"):CombineLatest(
+    FromUISettings("pet.useFixedColorForBackground", "pet.backgroundColor"))
+    :Map(function(quality, useFixedColor, fixedColor)
+        if useFixedColor then 
+          return fixedColor
+        end
+
+        local color = Color(ITEM_QUALITY_COLORS[quality - 1] or RED_FONT_COLOR)
+        return Color(color.r * darkenFactor, color.g * darkenFactor, color.b * darkenFactor, 0.73)
+    end)
+end
+
+function FromBorderColor()
+  return FromUIProperty("PetQuality"):CombineLatest(
+    FromUISettings("pet.useFixedColorForBorder", "pet.backgroundColor"))
+    :Map(function(quality, useFixedColor, fixedColor)
+        if useFixedColor then 
+          return fixedColor
+        end
+
+        local color = Color(ITEM_QUALITY_COLORS[quality - 1] or RED_FONT_COLOR)
+        return Color(color.r, color.g, color.b, 1)
+    end)
 end
 
 function FromPetNameTextColor()
   local darkenFactor = 1.0
 
-  return FromUIProperty("PetQuality"):Map(function(quality)
-    local color = Color(ITEM_QUALITY_COLORS[quality - 1] or RED_FONT_COLOR)
-    return Color(color.r * darkenFactor, color.g * darkenFactor, color.b * darkenFactor)
-  end) 
+  return FromUIProperty("PetQuality"):CombineLatest(
+    FromUISettings("pet.name.useFixedColor", "pet.name.fixedColor"))
+    :Map(function(quality, useFixedColor, fixedColor)
+        if useFixedColor then 
+          return fixedColor
+        end
+
+        local color = Color(ITEM_QUALITY_COLORS[quality - 1] or RED_FONT_COLOR)
+        return Color(color.r * darkenFactor, color.g * darkenFactor, color.b * darkenFactor, 0.73)
+    end)
 end
 -------------------------------------------------------------------------------
 --                                Styles                                     --
 -------------------------------------------------------------------------------
 Style.UpdateSkin("Default", {
   [PetView] = {
-    height = 28,
+    height                            = 28,
     registerForClicks                 = { "LeftButtonDown", "RightButtonDown" },
     backdrop                          = FromBackdrop(),
     showBackground                    = FromUISetting("pet.showBackground"),
     showBorder                        = FromUISetting("pet.showBorder"),
-    -- backdropColor                     = FromUISetting("pet.backgroundColor"),
     backdropColor                     = FromBackdropColor(),
-    backdropBorderColor               = FromUISetting("pet.borderColor"),
+    backdropBorderColor               = FromBorderColor(),
     borderSize                        = FromUISetting("pet.borderSize"),
 
     Icon = {
-      width = 20,
-      height = 20,
-      file = FromUIProperty("PetIcon"),
-      texCoords                     = { left = 0.07,  right = 0.93, top = 0.07, bottom = 0.93 } ,
-      subLevel            = 1,
-      location = {
-        Anchor("LEFT", 5, 0),
-      }
+      width                           = 20,
+      height                          = 20,
+      file                            = FromUIProperty("PetIcon"),
+      texCoords                       = { left = 0.07,  right = 0.93, top = 0.07, bottom = 0.93 } ,
+      subLevel                        = 1,
+      location                        = { Anchor("LEFT", 5, 0) }
     },
 
     SourceIcon = {
-      width = 14,
-      height = 14,
-      subLevel = 2,
-      file = FromUIProperty("SourceIcon"),
-      location = {
-        Anchor("BOTTOM", 0, 0, "Icon", "BOTTOMRIGHT")
-      }
+      width                           = 14,
+      height                          = 14,
+      subLevel                        = 2,
+      file                            = FromUIProperty("SourceIcon"),
+      location                        = { Anchor("BOTTOM", 0, 0, "Icon", "BOTTOMRIGHT") }
     },
 
     Name = {
-      text = FromUIProperty("PetName"),
-      textColor = FromPetNameTextColor(),
-      location = {
-        Anchor("TOP"),
-        Anchor("LEFT", 5, 0, "Icon", "RIGHT"),
-        Anchor("RIGHT"),
-        Anchor("BOTTOM")
-      }
+      text                            = FromUIProperty("PetName"),
+      textColor                       = FromPetNameTextColor(),
+      mediaFont                       = FromUISetting("pet.name.mediaFont"),
+      location                        = {
+                                        Anchor("TOP"),
+                                        Anchor("LEFT", 5, 0, "Icon", "RIGHT"),
+                                        Anchor("RIGHT"),
+                                        Anchor("BOTTOM")
+                                      }
     }
   },
 
   [PetListView] = {
-    viewClass = PetView,
-    indexed = false,
-    hideOwned = FromUISetting("pets.hideOwned"),
-    columns = FromUISetting("pets.columns")
+    viewClass                         = PetView,
+    indexed                           = false,
+    hideOwned                         = FromUISetting("pets.hideOwned"),
+    columns                           = FromUISetting("pets.columns")
   }
 })